backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
qrcode==7.4.2
pillow==10.2.0
reportlab==4.0.9